+++ tests/network/traffic_isolation_model.py
@@ -13,10 +13,10 @@ from typing import Dict, Mapping, Tuple
 APPROVED_HOSTS = (
     "gs-loc.apple.com",
     "gs-loc-cn.apple.com",
-    "gs-loc-corpa.apple.com",
-    "gs-loc.apple.com.cn",
+    "gsp-ssl.ls.apple.com",
     "bluedot.is.autonavi.com",
     "bluedot.is.autonavi.com.gds.alibabadns.com",
+    "gspe19-cn-ssl-ls-apple-com.v.aaplimg.com",
 )